requirements.txt
# C&C CRM API Dependencies
# FastAPI Backend Requirements

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
prisma==0.12.0
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP & API
httpx==0.25.2
requests==2.31.0

# File Upload & Storage
python-magic==0.4.27
Pillow==10.1.0

# Environment & Configuration
python-dotenv==1.0.0
pydantic==2.5.0

# Validation
# zod==0.0.1  # Commented out - not a valid Python package

# Date & Time
python-dateutil==2.8.2

# Logging
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Production
gunicorn==21.2.0

# Optional: AI/ML (for future features)
# scikit-learn==1.3.2
# pandas==2.1.4
# numpy==1.25.2

# Optional: Image Processing
# opencv-python==4.8.1.78

# Optional: PDF Generation
# reportlab==4.0.7

# Optional: Email
# fastapi-mail==1.4.1

# Optional: Caching
# redis==5.0.1

# Optional: Background Tasks
# celery==5.3.4 